requirements.txt
h5py==2.7.1
Keras>=2.1.4
numpy==1.14.1
opencv-python>=3.0
pandas>=0.22
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
tensorflow-gpu>=1.4